Kill use of context
This commit makes the transport synchronous which simplifies things for
for now. Introducing contexts was somewhat problematic, because we don't
necessarily have the ability to set reasonable deadlines for both read
and write calls. Until we can figure out a way to do that, let's just
not pretend that we can.