Some additional quick cleanup
We have a very infrequently hit race condition with the error
channels where it would be closed, but a goroutine would write
an error it after closing, causing a panic. This solves that by
simpling allowing the us to potentially leak a couple of
goroutines, errors, and channels.

This also kills this weird call out to netstat I added for
debugging once.