Shuffle store/etcd around, make e2e tests pass
Had to do a couple of things in this commit. I didn't like how
we were passing etcd to the Store objects, and so I got that out
of the way. Then, I also made the e2e tests pass by fixing a
couple of issues in the HTTP API.

The e2e tests now also don't talk to etcd directly--but instead
query the HTTP API.