No longer serialize messages going through WizardBus
This includes a number of changes around both the MessageBus and the
Transport (to facilitate increased test coverage in another PR). The
MessageBus interface now indicates that messages are passed across the
bus as an interface{}. Justification for this change is well-documented
in issue #187.

This commit also introduces the Transport interface. I've opened an
issue (#195) for the work.