Keepalived persistence, kepaalive store rewrite, monitor changes
The keepalive store now only stores failing keepalives namespaced
by backend. This allows a backend to recover the state of its failing
keepalives when it restarts. Agents may have reconnected to other
backends, and it guards against this by checking if an entity is
now passing its keepalive check before starting a new monitor.

The keepalive store now only stores failing keepalives. This should
save a lot of etcd throughput. We weren't using the passing
keepalives ever--as users would view the LastSeen field on the
entity directly.

The monitor now only updates keepalives if the entity is failing.