Remove time.sleep to avoid context switching
We were context switching once a millisecond instead of blocking
on IO and channel selects. Remove the defaults in the case
statements where this was happening and allow goroutines to be
parked indefinitely (allowing sensu agent and backend to have
their threads likewise parked by the kernel). This is like a
100x decrease (with linear multiplier) in context switching
per agent on the backend and a 100x decrease in context switching
in the agent.