Remove reconnects from the agent
The reconnect code in the Agent was never finished, and it introduces
undesirable complexity to its main run loop. Simplifying the main run
loop should allow us to make it easier to test the agent and make its
behavior more consistent. Cause the agent to crash when we can't
connect to the backend, allowing process managers to restart us. This
has the added bonus of allowing people to monitor a repeatedly crashing
sensu-agent process with their process manager.