Don't try to gracefully shutdown the HTTP server
I'd rather we didn't half-heartedly try to shutdown the HTTP server
in the agent. If we _look_ like we gracefully shutdown in the code,
that's going to make people assume that we done, but we can't.
The stopping channel is going to shutdown the send/receive pumps.
This means that if we _are_ randomly able to gracefully shutdown
due to scheduling favoring the HTTP Server goroutine, then we have
a race where we have data in the send/recvq that's been accepted
that will never be sent.

This is already the case, but I'd rather we didn't exacerbate the
problem. If we want to have a contract of "if we have received a
message, we will ensure it is sent," then that is a much larger
feature.