Make agent.Stop() block until listeners stop
In an attempt to fix our racy tests where listening on the agent
sockets or potentially the API can cause tests to fail, we'll make
the Stop() method block until everything has shutdown cleanly.

Go back to having separate goroutines monitoring the stopping
channel in order to shutdown the HTTP API and the TCP Listener.
Everything else is pretty straightforward.