Accept subscriptions in comma-separated or list form.
Subscriptions as designed were meant to be comma-separated in
both the pflag and yaml configuration contexts. However, this is
somewhat confusing for yaml authors, since lists of things
are typically specified as such in yaml.

This change allows specifying subscriptions in both forms,
preserving the existing behaviour for comma-separated inputs, and
also supporting lists.

Refs #515