Add package dynamic.
Package dynamic can be used to with json.Marshaler and
json.Unmarshaler to create types that can support arbitrary
custom attributes.

The package achieves this by parsing json messages with
the json-iterator package in a streaming fashion, and mapping the
results onto dynamically created types with package reflect.

Package dynamic also implements GetField, which can be used to
implement the govaluate.Parameters interface.