Don't delete silenceds in bulk.
Deleting silenceds in bulk is not a functional requirement, and is
potentially more destructive than the user intended. Simplify the
Destroy method for the silenced controller, and remove the ability
to delete silenceds in bulk.

Also adds a convenience function to the types package for computing
the ID of a silenced.

Fixes #640