Add check subdue command to the CLI.
This commit adds the ability to subdue checks via the CLI. Users can
send subdue definitions in JSON format by stdin or by file.

This commit also adds validation of Check.Config.Subdue, by checking
that all subdue definitions are well formed. It does this by copying
the definitions in the util/time package and making methods out of
them. This was done to avoid import cycle issues between the two
packages.

If reviewers like the way this code is moved to the types package,
then a followup ticket should be filed to remove the util/time
package.

The local client will attempt to reconcile locations that exist in
the time specifiers. If no location is specified, time.Local is
used and the offset is subtracted from UTC. The client sends all
times to the server in UTC.

Signed-off-by: Eric Chlebek <eric@sensu.io>