Use UTC for InWindow y/m/d calculation
time.InWindow takes a time.Time, and uses it to get the year,
month, and date without first ensuring that it is in UTC. Begin and End
times for the window were calculated based on UTC. In cases where the
time passed in was local time, begin/end would be in advance of the
current time, causing InWindows to return false. time.InWindow has been
updated to get the y/m/d from the current UTC time so the begin and end
time comparison will be more accurate.

Signed-off-by: Mercedes Coyle <mercedes@sensu.io>