Support special characters in resource names.
This commit supports resources with names that would need to be
percent-encoded in URLs in order to match properly with our mux,
for example 'foo/bar'. To this end, all resource names are
encoded before being sent to the backend, and all resource names
are decoded in the backend before being looked up.

In doing this work, a bug was also found where mutator destroy was
not functioning correctly. This bug has also been fixed.

The default 404 response now writes a JSON-encoded response body,
in order to match the response generated by the rest of the API.

gorilla/mux has been upgraded, just because.

Signed-off-by: Eric Chlebek <eric@sensu.io>