Refactor the store, queue and ring.
This is a significant refactoring that aims to accomplish several
things.

* Don't combine the Store interface with Queue and Ring interfaces.

* Ensure that Queue and Ring are decoupled and represented in the
types package.

* Provide mock and fake implementations of Queue and Ring.

* Don't export internal state of various daemon types. Instead,
create Newers that handle setting the daemons' internal state.
This is more idiomatic and also reduces the chance of needing to
break the API, by having fewer symbols exported.

* Create an option interface for all the daemon Newers, so that
the API won't need to break as we add or deprecate functionality.

* Use consistent terminology and APIs between queue and ring
packages for getters.

* Define types.QueueItem, and have Queues implement it for Dequeue.

Signed-off-by: Eric Chlebek <eric@sensu.io>