Move Hooks and Silenced into Check. (#1149)
* Add Hooks and Silenced to Check protobuf.
* Deprecate, but do not remove, Event Hooks and Silenced.

When events are unmarshaled, move any hooks and silenceds out of
the event and into its check. Nil-out the hooks and silenceds in
the event itself, so that when it is marshaled, the hooks and
silenceds are put into the proper place.

In the future, we plan to remove Hooks and Silenced from the Event
protobuf. This should happen before we start depending on proto
marshaling and unmarshaling.

Signed-off-by: Eric Chlebek <eric@sensu.io>