Add dynamic resource resolution.
* Add type Resource, which is an interface that specs URIPath.
* Add function ResolveResource, which can deliver a zero-valued
type which implements Resource, given a valid Resource name.
* Add type Wrapper, which users can supply in json format. The
wrapper type contains a type name and a spec. When Wrapper is
unmarshaled, ResolveResource is called to discover its concrete
type.

ResolveResource is implemented using go generate and a go program
that uses godoc to discover all of the exported types in the types
package.

Signed-off-by: Eric Chlebek <eric@sensu.io>