Add sensuctl create command.
The sensuctl create command can create arbitrary Sensu resources
by parsing one or more Sensu resource definitions.

As a convenience, users can omit organization and environment in
the resources, and the command-line values for these fields will
be used instead.

This commit fixes a few issues with the work in the types package
that was done in the previous commit.

Signed-off-by: Eric Chlebek <eric@sensu.io>