Fix a deadlock in the monitor.
The deadlock only occurs under load; in testing, it took around
5500 agent sessions to expose the problem.

On reset, the monitor acquires a mutex and then performs a channel
send. However, if another select case in the monitor is trying to
acquire the mutex, then the channel send will never occur.

Adding a buffer to the resetChan fixes the problem.

Signed-off-by: Eric Chlebek <eric@sensu.io>