Add ExtensionExecutor.
ExtensionExecutor is meant to be used by components that retrieve
Extensions from the store, and need to execute them with gRPC.

ExtensionExecutor maintains a package-level cache of gRPC client
connections, so that repeatedly dialing gRPC services can be
avoided. It also wraps and unwraps the Request types that are
required by the gRPC service and client.

Signed-off-by: Eric Chlebek <eric@sensu.io>