Add e2e test for extensions.
This commit adds an end-to-end test for extensions. It also removes
the gRPC connection cache that extensions relies on, as it was
discovered to not be working properly in the course of testing.

As of now, extensions are working, but each extension request will
establish a new gRPC connection. So extensions are likely not
performant yet.

Signed-off-by: Eric Chlebek <eric@sensu.io>