Replace found with react-router
`react-router` is a better fit for this project. It is a more familiar tool within the community that will empower more people to contribute and provide value to the project.

Since migrating from `relay` to `apollo` our router is no longer coupled to the GraphQL client implementation, allowing us to make this change.

Signed-off-by: Neal Granger <neal@nealg.com>