Optimize check marshaling and add a benchmark for subPump (#1306)
This commit significantly reduces the amount of time spent
marshaling json in agentd. It replaces stdlib encoding/json with
json-iterator, and it avoids using dynamic.Marshal when there are no
dynamic attributes to be marshaled.

Profiling performed during load testing shows that json encoding
and decoding can comprise a substantial portion of time spent in
the backend. This is most evident when there are many thousands of
checks being executed per second.

The benchmark results below are for a single iteration of subPump.
subPump is part of agentd; it receives messages on a channel, and
forwards them along to the agent. Profiling shows that the majority
of time in subPump is spent dealing with json rendering. For every
check that an agent executes, it must process a single subPump
iteration.

Before:
26323 ns/op   16770 B/op  149 allocs/op

After:
5863  ns/op   3822  B/op  40  allocs/op

Note that memory allocations are significantly reduced as well,
reducing GC pressure by over a factor of 2.

Signed-off-by: Eric Chlebek <eric@sensu.io>