Populate Issued field for Check and CheckRequest.
This commit fixes a bug where the Issued field of CheckRequest and
Check were never populated.

Issued is the unix time at which the CheckRequest was issued.

The Issued field had to be added to CheckRequest, as it was
previously missing.

We also found that json-marshaling the Check would result in a
missing value for issued; that has also been fixed.

The agent now correctly proxies the CheckRequest Issued field into
the Check Issued field.

Closes #1324

Signed-off-by: Eric Chlebek <eric@sensu.io>