[Web UI] Use apollo-link-state for client-side auth state (#1352)
apollo-link-state represents an opportunity to use GraphQL to manage local client state as an alternative to something like redux.

This change introduces patterns for managing local state with apollo and replaces the existing singleton auth token storage mechanism. The new auth token storage as part of the GraphQL client can be interacted with through the same query and mutation interface as the remote GraphQL endpoint. This makes consuming and modifying any state within components much simpler and more opaque.