Allow nullable flap thresholds.
This commit allows users to indicate negative presence of a flap
threshold by supporting 'null' as a valid JSON value. It does this
by defining a proto message called FlapThreshold, which can then
be assigned nil.

This is necessary because in proto3, scalar values cannot be
pointers in generated code.

In GraphQL, a value of -1 will indicate the absence of a flap
threshold value.

Signed-off-by: Eric Chlebek <eric@sensu.io>