[GraphQL] Add offset pagination to large connections
Existing pagination works well with infinite scrolling and when a
resource's position in a sorted set is immutable (eg. when sorted by the
date created). However, when implementing simple pagination with
multiple sorting options the product of page quickly becomes
unintuitive.

As such this PR switch the checks, entities and events connections to
use a simplistic offset-based pagination solution.

Existing supporting code has been removed but may re-introduced if we
need an event stream or activity feed in the future.

Signed-off-by: James Phillips <jamesdphillips@gmail.com>