Improve display and ergonomics of silenced entries. (#1447)
This commit adds a 'Begin' time to silenced entries so that users
will be able to recognize which entries have already begun and
which haven't. Also, expiry time is computed as an offset from
begin, if the silenced entry hasn't begun yet.

The commit also improves the ergonomics of using the interactive
mode for creating silenced entries, by defaulting the silenced
begin time to 'now'. (Previously, the default was 0, which was
both confusing, and also interpreted as the start of the unix
epoch.)

Signed-off-by: Eric Chlebek <eric@sensu.io>