Refactor end-to-end tests.
This commit significantly changes how our end-to-end tests work.
Instead of each test spinning up its own backend, the tests all
share a single backend, with two exceptions.

The reason why we want to constrain the number of backends running
is that allocating ports to these backends is somewhat tricky, and
with a static number of backends, ports can be allocated
statically.

Initially, it was planned that all tests would run on the same
backend, but that didn't pan out for the agent reconnection tests
and the RBAC tests.

This commit also removes the HTTP client in favour of sensuctl.
sensuctl is now used for all e2e tests.

Tests are isolated from one another via RBAC - each test executes
within its own organization and environment.

It remains to be seen if this strategy will lead to a more stable
e2e testing environment. I had a lot of issues in implementing this
work.

Signed-off-by: Eric Chlebek <eric@sensu.io>