Fix a hook encoding bug related to embedding.
Due to the protobuf compiler generating a JSON tag, JSON encoding
and decoding of Hooks was broken. When an embedded struct has a
JSON tag defined, the JSON encoder behaves differently, and does
not treat the embedded type as embedded.

This commit sets the tag to "", which works around this.

Signed-off-by: Eric Chlebek <eric@sensu.io>