Fix a bug where namespace flags were removed.
This commit reverts a regression caused by a previous commit,
d6f761f746ef94ab2cc8d9649a171ec0fa4274e6, that attempted to
hide the organization and environment flags only for the create
command.

That didn't work, and hid the flags for all commands. Undo what
the bad commit did, and issue a warning if the --organization and
--environment flags are set on create.

Signed-off-by: Eric Chlebek <eric@sensu.io>