Remove apollo-link-error
apollo-link-error doesn't serve our desired error handling pattern. We want any query or mutation consumer to be able to initially respond to any error state, and only handle an error centrally if the consumer didn't swallow the error. Unfortunately apollo-link-error intercepts errors before they are provided to the consumer and is called with any error regardless of the consumer behavior. It is better to rely strictly on consumer handling, although this does mean we have to manually throw errors in queries and mutations where the default apollo behavior is to silence them.

Signed-off-by: Neal Granger <neal@nealg.com>