[Web UI] Trap uncaught errors and promise rejections (#1543)
* Trap uncaught errors and promise rejections

Set up a global error detection system connected where any unhandled exception or promise rejection will trigger a global error state.

This change implements the detection system. Detected errors are currently logged and not handled in any way (coming with #1360).

* Create FetchError class and fetch util

When using `utils/fetch.js`, any issues encountered during a fetch call (network/coors errors or 400/500 http status) are wrapped in a FetchError instance for consistent and reliable error matching logic in downstream handlers.

* Remove apollo-link-error

apollo-link-error doesn't serve our desired error handling pattern. We want any query or mutation consumer to be able to initially respond to any error state, and only handle an error centrally if the consumer didn't swallow the error. Unfortunately apollo-link-error intercepts errors before they are provided to the consumer and is called with any error regardless of the consumer behavior. It is better to rely strictly on consumer handling, although this does mean we have to manually throw errors in queries and mutations where the default apollo behavior is to silence them.

Signed-off-by: Neal Granger <neal@nealg.com>