[Web UI] Prevent cascading errors when token refresh fails (#1555)
* Rewrite authLink to abort operation on auth fail

In the event that an auth token refresh fails while processing a mutation or query, abort the operation with the failed token request error. There is no sense in continuing to attempt the operation if authentication has failed. This prevents a cascade of additional failed attempts with the API.

* Memoize auth API request functions

This change disallows multiple parallel requests to the auth API with the same data. If a token refresh request is fired while an identical one is in-flight, the `refreshTokens` function will resolve the current pending fetch for both requests.

This is achieved through a promise-aware memoization utility that keeps a cache of current results of a given promise-returning function according to a configured key creator function.

Signed-off-by: Neal Granger <neal@nealg.com>