Fix an aliasing bug in ResolveResource.
This commit fixes a bug where the same pointer is returned from
ResolveResource every time. This causes havoc with functions that
expect the value coming from ResolveResource to be a new resource
every time.

The effect of this was that unmarshalling tended to simply not work
properly.

This commit uses the reflect package to copy the type from the
type map, providing a new Resource with every invocation.

Signed-off-by: Eric Chlebek <eric@sensu.io>