Fixed a bug where environments could not be deleted.
This commit fixes a bug where environments could not be deleted if
the organization they reside in has one or more assets.

Assets do not exist within environments, so there is no reason to
consider them when deleting environments.

This commit also stops the environment store from reporting an
error when an environment that doesn't exist is deleted.

Closes #1567

Signed-off-by: Eric Chlebek <eric@sensu.io>