[Web UI] Add ESLint rule 'curly'
JavaScript allows the omission of curly braces when a block contains
only one statement. However, it is considered by many to be best
practice to never omit curly braces around blocks, even when they are
optional, because it can lead to bugs and reduces code clarity.

Also, makes it more difficult to set breakpoints.

Signed-off-by: James Phillips <jamesdphillips@gmail.com>