Pass SENSU_BUILD_ITERATION to packaging container (#1641)
Signed-off-by: Terrance Kennedy <terrancerkennedy@gmail.com>

## What is this change?

Passes the `SENSU_BUILD_ITERATION` environment variable into the sensu-go-build docker container when deploying packages.

## Why is this change necessary?

Sensu nightly builds were broken on Travis because the version command could not determine the build iteration.

## Does your change need a Changelog entry?

Not this time.

## Do you need clarification on anything?

I'm good.


## Were there any complications while making this change?

I noticed that the package_cloud gem was being installed on the host, which is unnecessary because it's only installed/used inside the sensu-go-build docker container. Builds should be a few seconds faster now too.