Connect silence entry form to API
This is fully functional for singular and bulk creation, but has some rough edges with server error handling.

Currently no two silencing entries can have the same check and subscription (due to how they are indexed internally). This manifests as an error on the mutation that we are displaying as a field validation error within the form. This approach isn't ideal but at least it gets everything working without any expected errors going unhandled.

This work avoids any success notification, the form dialog simply closes when no submit errors are encountered. The intention is to create a success flash or toast notification of some sort to provide positive user feedback in the future.

Signed-off-by: Neal Granger <neal@nealg.com>