Refactor monitors to use etcd
Monitors are based on a leased key in etcd. A watcher waits for a DELETE
event and fires a failure handler when that event is seen. If a key has
a PUT event, its watcher is terminated to avoid multiple watchers on the
same key.

Co-authored by: Mercedes Coyle <mercedes@sensu.io>
Co-authored by: Eric Chlebek eric@sensu.io

Signed-off-by: Mercedes Coyle <mercedes@sensu.io>