Add nuance to govaluate error handling. (#1690)
Add two error types to the eval package, which allows users of the
package to distinguish between syntax errors and type errors.

In the asset manager, if errors are not SyntaxErrors or TypeErrors,
then only issue a debug log entry. Otherwise issue an error log
entry. Do the same thing for proxy check entities.

Closes #1512

Signed-off-by: Eric Chlebek <eric@sensu.io>