Don't reset check occurrences to 1 on check resolution.
Check occurrences are the number of times a check has had the same
status, in the case where checks are failing.

When checks are resolved, the occurrences field will stop
receiving updates. Occurrences only apply to failing status codes.

This is somewhat unintuitive, but it makes 2.x compatible with 1.x
in this respect.

Closes #1624

Signed-off-by: Eric Chlebek <eric@sensu.io>