Fix a goroutine leak in Ring.Next.
Next creates a goroutine that watches an etcd key for deletes.
After Next returns, the goroutine is supposed to be shut down, but
isn't, because the context it is using to determine cancellation is
the wrong one.

This commit refactors the watcher into a separate function, and
passes the correct context into it.

Closes #1746

Signed-off-by: Eric Chlebek <eric@sensu.io>