Upgrade gometalinter to v2 (#1741)
Signed-off-by: Terrance Kennedy <terrancerkennedy@gmail.com>

## What is this change?

Upgrade gometalinter to v2.


## Why is this change necessary?

I upgraded gometalinter in sensu-enterprise-go after submitting a patch for v1 and learning that it is no longer supported.

## Does your change need a Changelog entry?

Added!

## Do you need clarification on anything?

The `install_deps` function inside build.sh installs gocov and golint, but neither appear to be used anywhere else in the build system. Can they be removed?


## Were there any complications while making this change?

The gometalinter command's invocation did not need to be changed, making this an easy PR.

## Have you reviewed and updated the documentation for this change? Is new documentation required?

No docs required!