Re-work ring for round-robin check execution.
This commit re-works the ring data structure after testing revealed
that round robin check execution in the presence of multiple
backends could be too frequent.

The ring is now administered by only a single writer. Backends race
to obtain a mutex on each ring, and the backend that obtains the
lock is responsible for administering the ring. The backends are
apprised of ring events via etcd watchers.

Round-robin check execution now works properly in the presence of
multiple backends.

This commit also removes the un-used Peek method of the ring.

Signed-off-by: Eric Chlebek <eric@sensu.io>