Add sensuctl cluster member-update command.
The sensuctl cluster member-update command can be used to modify
the peer URLs of an existing cluster member.

This commit also modifies the output of the list command to print
cluster member IDs in hex format, and also modifies the API to
expect ID parameters to be in hex format. This reflects the
behaviour of the etcdctl commands.

Signed-off-by: Eric Chlebek <eric@sensu.io>