Use crock to mock out time in schedulerd.
This commit abstracts the time package in schedulerd by using
timeproxy. In tests, the time is mocked out with crock time.

The result is that these tests no longer need to sleep in order
to verify that scheduling is working properly.

Test execution has been sped up by a factor of 10-30x as a result.

Signed-off-by: Eric Chlebek <eric@sensu.io>