Add a deadline to extension methods.
This commit adds a hardcoded 5 second deadline to extension
methods. It has been observed that extensions can cause a lockup
in the backend, and this commit is designed to mitigate that.

More investigation is needed into why extensions are locking up in
the first place, but this is a starting point that makes the
situation less critical.

Signed-off-by: Eric Chlebek <eric@sensu.io>