Re-work queue to support one-to-many messaging.
This commit modifies the queue so that each backend can consume
messages from it - also known as fan-out.

As an unrelated refactoring change, this commit also factors out
the sequencing method from the ring and places it in the etcd
package.

I initially intended to use the ring sequencing method for the
queue, but later remembered that the queue uses timestamps, as it
relies on the timestamp value to compute queue item freshness.

During the course of this work, I discovered that the queue is not
safe to use with multiple producers and consumers for the same
backend. To support this use case, I believe we would need to
use etcd's concurrency.STM datatype.

However, we do not currently have this use case, so I think such
a feature is not needed. Therefore, the test case that tests
concurrent Enqueue and Dequeue has been removed, as it can
sometimes fail.

I have tested this commit with clustered sensu, and have observed
that ad-hoc checks now function as expected.

Closes #1904

Signed-off-by: Eric Chlebek <eric@sensu.io>