Re-work queue to support one-to-many messaging. (#1942)
* Re-work queue to support one-to-many messaging.

This commit modifies the queue so that each backend can consume
messages from it - also known as fan-out.

As an unrelated refactoring change, this commit also factors out
the sequencing method from the ring and places it in the etcd
package.

I initially intended to use the ring sequencing method for the
queue, but later remembered that the queue uses timestamps, as it
relies on the timestamp value to compute queue item freshness.

During the course of this work, I discovered that the queue is not
safe to use with multiple producers and consumers for the same
backend. To support this use case, I believe we would need to
use etcd's concurrency.STM datatype.

However, we do not currently have this use case, so I think such
a feature is not needed. Therefore, the test case that tests
concurrent Enqueue and Dequeue has been removed, as it can
sometimes fail.

I have tested this commit with clustered sensu, and have observed
that ad-hoc checks now function as expected.

Closes #1904

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Fix a bug where splitting queue keys on '/' was invalid.

When raw bytes are used for etcd keys, a small number of the keys
will incidentally contain '/' in surprising places. Therefore,
don't use raw bytes, but rather encode them to hexadecimal.

Also add the parallel dequeue test back, since this change makes
this test reliably pass again.

Remove a redundant error check in the ring.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Fix lint, don't build queue tests for -race.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Minor code review fixes.

Signed-off-by: Eric Chlebek <eric@sensu.io>