Support using sensu-backend with external etcd instances
This commit adds the --no-embed flag to the backend. When this flag
is specified, sensu-backend will dial whatever URLs are specified
in the --listen-client-urls flag, and not start up etcd.

This change required some refactoring of the queue package, which
depended on the cluster member ID for operation. This ID has been
replaced by a client lease ID which is created on backend startup.

As a result, this made queue garbage collection easier to
guarantee, because now all queue keys are created with a lease.
The test that tested queue garbage collection has been removed,
because the logic that removes the keys is no longer part of the
queue package.

Testing was done by starting up an external etcd instance and
pointing the backend at it. No issues were observed, and the etcd
instance was correctly populated on startup.

Signed-off-by: Eric Chlebek <eric@sensu.io>