Retry lease acquisition in a loop for backend ID.
This commit adds retry behaviour to the backend ID feature. As
such, an interface needs to be passed into the queue library.

The queue library now no longer needs to use a keepalive - this is
now handled by the backend lease keepalive.

The move from cluster member IDs to lease IDs means that the queue
durability is impacted. Client mailboxes can no longer survive
extended network outages or process restarts.

It may be necessary to revisit this in the future to increase
durability, but for our current use case, this is fine, as ad-hoc
check requests are fairly ephemeral.

Signed-off-by: Eric Chlebek <eric@sensu.io>