Support using sensu-backend with external etcd instances (#2037)
* Support using sensu-backend with external etcd instances

This commit adds the --no-embed flag to the backend. When this flag
is specified, sensu-backend will dial whatever URLs are specified
in the --listen-client-urls flag, and not start up etcd.

This change required some refactoring of the queue package, which
depended on the cluster member ID for operation. This ID has been
replaced by a client lease ID which is created on backend startup.

As a result, this made queue garbage collection easier to
guarantee, because now all queue keys are created with a lease.
The test that tested queue garbage collection has been removed,
because the logic that removes the keys is no longer part of the
queue package.

Testing was done by starting up an external etcd instance and
pointing the backend at it. No issues were observed, and the etcd
instance was correctly populated on startup.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Moved the sensu-agent binary to ./cmd/sensu-agent. (#2046)

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Support using sensu-backend with external etcd instances

This commit adds the --no-embed flag to the backend. When this flag
is specified, sensu-backend will dial whatever URLs are specified
in the --listen-client-urls flag, and not start up etcd.

This change required some refactoring of the queue package, which
depended on the cluster member ID for operation. This ID has been
replaced by a client lease ID which is created on backend startup.

As a result, this made queue garbage collection easier to
guarantee, because now all queue keys are created with a lease.
The test that tested queue garbage collection has been removed,
because the logic that removes the keys is no longer part of the
queue package.

Testing was done by starting up an external etcd instance and
pointing the backend at it. No issues were observed, and the etcd
instance was correctly populated on startup.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Disable auto text wrapping in sensuctl (#2047)

Disable auto text wrapping in sensuctl

Signed-off-by: Nikki Attea <nikki@sensu.io>

* Retry lease acquisition in a loop for backend ID.

This commit adds retry behaviour to the backend ID feature. As
such, an interface needs to be passed into the queue library.

The queue library now no longer needs to use a keepalive - this is
now handled by the backend lease keepalive.

The move from cluster member IDs to lease IDs means that the queue
durability is impacted. Client mailboxes can no longer survive
extended network outages or process restarts.

It may be necessary to revisit this in the future to increase
durability, but for our current use case, this is fine, as ad-hoc
check requests are fairly ephemeral.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Remove non-essential queue functionality.

The queue can no longer support durability, since we can no longer
rely on a persistent backend ID. However, durability is not
currently a requirement for adhoc check requests. So, features
pertaining to durability have now been removed.

Durability can be supported with additional engineering effort,
but since nothing requires that right now, it's preferable to do
the simpler thing.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Mention that the queue is non-durable.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Tired: --no-embed <> Wired: --no-embed-etcd

Signed-off-by: Eric Chlebek <eric@sensu.io>