Rewrite RequestInfo middleware using gorilla
This simplifies the middleware and lets gorilla take care of parsing the
path.

The flip side is that since the gorilla router is fairly strict in its
matches, we lose the "best effort filling in" property of the
RequestInfo middleware previously defined.

We could abuse the router to have the previous behaviour again, where
even the group on its own could be extracted, but I don't think it's
worth it.

Signed-off-by: Cyril Cressent <cyril.cressent@sensu.io>