Add freeze-api tool. (#2187)
Add freeze-api tool.

This commit adds a tool, freeze-api, that can be used to freeze
an internal sensu API to a versioned package. It is intended
to handle package renaming, type registration and conversions.

The gen-register tool can be used to register types in a central
registry. Types can then be Resolved with a TypeMeta key.

The type registry will be used to dynamically resolve types at
runtime, notably for use with tools like sensuctl create.

A conversion utility has been added to the freeze-api command. This feature
generates type converters between the package being frozen and
the package being created.

For now, only pointer conversions are supported. In the future,
a more featureful conversion interface will be required.

Signed-off-by: Eric Chlebek <eric@sensu.io>