Add support for +resource-name.
All package-level types that are frozen will now be given a
ResourceName method. By default, the method will return an empty
string.

By annotating a type with the +resource-name option, users can set
what the output of the ResourceName method will be.

This commit adds such annotations to the rbac package. This was
used for testing the functionality, as to give the rbac package
types usable resource names.

Signed-off-by: Eric Chlebek <eric@sensu.io>