Add support for +resource-name. (#2214)
* Add support for +resource-name.

All package-level types that are frozen will now be given a
ResourceName method. By default, the method will return an empty
string.

By annotating a type with the +resource-name option, users can set
what the output of the ResourceName method will be.

This commit adds such annotations to the rbac package. This was
used for testing the functionality, as to give the rbac package
types usable resource names.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Add v1alpha1 meta package.

This commit adds the first version of the meta package. It also
fixes some bugs in the codegen tool - meta types were not
recognized as Kinds, because the generator assumed that all Kinds
would contain meta.TypeMeta!

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Fix some codegen bugs. Add lower-case kind lookup.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Fix some gofmt issues.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Add tests for converters.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Avoid some more gofmt lint.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* More gofmt noodling.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Address PR feedback.

* Improve verbiage in tests and elsewhere.
* Run gofmt and go generate as part of freeze-api.
* Prefix resource-name tag with freeze-api.
* Make code a bit DRYer.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Move conversion functions into internal packages.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Add unit tests for conversion package.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Add conversion registry.

Gofmt the whole project on freeze-api.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Use lowercase instead of camelcase in the registry.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Add update-protobuf.sh to freeze-api.

Signed-off-by: Eric Chlebek <eric@sensu.io>