Rework agent connection handling and send/recv loops
The send and receive loops in the agent were a tangled web of
multiple goroutines spread across numerous functions making it
more difficult to understand than it should be. This collapses
all of the connection handling into two goroutines and two
functions while providing a minimal amount of indirection for
the API's /healthz endpoint to still be able to read the state
of the agent's connection without requiring access to the
underlying transport interface. This also removes reconnection
logic from the transport which was ultimately implemented partially
in the agent and also in the transport.

Signed-off-by: Greg Poirier <greg.istehbest@gmail.com>