Merge the assets and check environments
The values of the PATH, CPATH and LD_LIBRARY_PATH variables are merged,
with a preference given to the values held in the check's environment.
That is, paths coming from a check's environment take precedence, but do
not overwrite paths coming from the assets' environment.

Any other variable is merged, with conflicts resolved in favor of the
check environment.

Signed-off-by: Cyril Cressent <cyril.cressent@sensu.io>