Set up flow type checker
Add configuration to the repo to support running flow. This does not introduce any changes to existing tests and will not affect CI. Flow is only active when run manually in development. Files without a `// @flow` pragma comment are fully ignored, allowing us to enable type checking on a per-file basis.

Signed-off-by: Neal Granger <neal@nealg.com>