Make RBAC list commands work similarly as the rest of the CLI
The RBAC list sub-command for role, role-binding, cluster-role and
cluster-role-binding returned an error not found message if there was no
resource to list. While this is correct behaviour, it is not aligned
with the rest of the list sub-commands, for which an empty result set is
returned.

This commit makes it so the RBAC list sub-commands follow the same
behaviour as all the other list sub-commands by _ignoring_ the not found
error.

Signed-off-by: Cyril Cressent <cyril.cressent@sensu.io>