Move all Resource types to core/v2 package
This commit moves all of Sensu's Resource types to the core/v2
package. This is part of a larger effort to version Sensu's API.

Type aliases have been added in the types package, to avoid
breaking consumers of the types package.

A side effect of this change is that mock.AnythingOfType no longer
works, as it relies on string matching. To work around this
problem, calls to mock.AnythingOfType have been replaced wherever
the call used a type from the types package.

The types.Wrapper machinery has been updated to rely on the v2
package. In the future, the types.Wrapper machinery should be able
to dynamically determine which package to look for types in.

Signed-off-by: Eric Chlebek <eric@sensu.io>