Fix Relocation sink interface
The `Sink` feature of relocation enables declaratively defining components to be rendered elsewhere. Compared to the currently used imperative interface, where a new component instance is spawned and handed-off, a declaratively defined component using `Sink` is coupled to it's parent component's lifecycle - behaving in a manor very similar to `React.Portal`.

Signed-off-by: Neal Granger <neal@nealg.com>