Return empty slice instead of nil for assets
When querying for a list of assets, instead of returning nil when there
are none to be returned, we return an empty slice of results instead.

This is to be in line with the behaviour of all the other object stores
and ensure a consistent behaviour across all the types.

Signed-off-by: Cyril Cressent <cyril.cressent@sensu.io>