Add blocking behaviour for Next
This commit changes the behaviour of Next, so that ring clients
synchronize with each other by keeping counters that are compared
with a counter for the ring.

When clients advance the state of the ring, they also increment
the counter.

When a client is found to be lagging behind the
counter on Next, their counter is set to the value of the ring's
counter, and ErrNotOwner is returned.

When a client is up-to-date with the ring's counter, they must
block until a delete event is observed on the head on the ring,
before returning ErrNotOwner.

This ensures that timing of Next calls does not matter - for a
given ordering of Next calls by a set of clients, the same results
will be delivered every time.

Signed-off-by: Eric Chlebek <eric@sensu.io>